+++ fiboa_cli/convert_utils.py
@@ -27,7 +27,7 @@ def convert(
         attribution = None,
         store_collection = False,
         license = None,
-        compression = "brotli",
+        compression = None,
         **kwargs):
     """
     Converts a German field boundary datasets to fiboa.